[PLAT-761] Fix treasury too low balance (#249)
* feat(aaa): remove treasury migration to do it manually

* feat(aaa): revert back treasury account id

* test(aaa): update for MockAccountId change

+++ pallets/ajuna-awesome-avatars/src/mock.rs
@@ -32,7 +32,7 @@ use sp_runtime::{
 	traits::{BlakeTwo256, IdentityLookup},
 };
 
-pub type MockAccountId = u32;
+pub type MockAccountId = u128;
 pub type MockBlockNumber = u64;
 pub type MockBalance = u64;
 pub type MockIndex = u64;
